move Message entity to domain module
