+++ library-compose/src/main/java/com/github/lexleontiev/chatexample/app/ComposeActivity.kt
@@ -3,11 +3,6 @@ package com.github.lexleontiev.chatexample.app
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Scaffold
 import androidx.compose.material.Text
@@ -17,8 +12,8 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.tooling.preview.Preview
 import com.github.lexleontiev.chatexample.app.ui.components.ChatScreen
-import com.github.lexleontiev.chatexample.app.ui.components.Message
-import com.github.lexleontiev.chatexample.app.ui.components.Message.Companion.mockList
+import com.github.lexleontiev.chatexample.library.Message
+
 
 class ComposeActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -43,7 +38,7 @@ fun AppMain() {
             backgroundColor = MaterialTheme.colors.background
         ) {
             val messages = remember { mutableStateOf(mutableListOf<Message>().apply {
-                addAll(mockList())
+                addAll(Message.mockList())
             }) }
             ChatScreen(
                 messages = messages,
